+++ Sources/bl-speech-recognizer/Utils/BLSpeechSynthesizer.swift
@@ -8,6 +8,12 @@
 
 import AVFoundation
 
+public struct Voice: Hashable {
+  public var language: String
+  public var identifier: String
+  public var name: String
+}
+
 protocol BLSpeechSynthesizerDelegate: AnyObject {
   func synthesizerStarted()
   func synthesizerFinished()
@@ -18,14 +24,19 @@ class BLSpeechSynthesizer: NSObject, @unchecked Sendable {
   weak var delegate: BLSpeechSynthesizerDelegate?
   private var buffer = BLResponseStringBuffer(minLength: 10)
   private var isFinished = false
-  private var language: String!
+  private var voice: AVSpeechSynthesisVoice!
 
   var isSpeaking: Bool {
     return synthesizer.isSpeaking
   }
 
   init(language: String) {
-    self.language = language
+    self.voice = AVSpeechSynthesisVoice(language: language)
+  }
+  
+  init(voice: Voice) {
+    let voiceIdentifier = voice.identifier
+    self.voice = AVSpeechSynthesisVoice(identifier: voiceIdentifier)
   }
 
   func stop() {
@@ -38,17 +49,24 @@ class BLSpeechSynthesizer: NSObject, @unchecked Sendable {
     buffer.onMessageReceived(text: str)
     internalSpeak()
   }
+    
+  static func availableVoices() -> [Voice] {
+    return AVSpeechSynthesisVoice.speechVoices().map { voice in
+      Voice(language: voice.language,
+            identifier: voice.identifier,
+            name: voice.name)
+    }
+  }
 
   private func internalSpeak() {
     guard synthesizer.isSpeaking == false else { return }
     buffer.flush(all: isFinished) { text in
       let utterance = AVSpeechUtterance(string: text)
-      utterance.voice = AVSpeechSynthesisVoice(language: language)
+      utterance.voice = self.voice
       synthesizer.delegate = self
       synthesizer.speak(utterance)
     }
   }
-
 }
 
 extension BLSpeechSynthesizer: AVSpeechSynthesizerDelegate {